refactor: refactored code according to best practices
