+++ openedx/core/djangoapps/notifications/base_notification.py
@@ -7,6 +7,7 @@ from .email_notifications import EmailCadence
 from common.djangoapps.student.roles import CourseInstructorRole, CourseStaffRole
 from .utils import find_app_in_normalized_apps, find_pref_in_normalized_prefs
 from ..django_comment_common.models import FORUM_ROLE_ADMINISTRATOR, FORUM_ROLE_MODERATOR, FORUM_ROLE_COMMUNITY_TA
+from .notification_content import get_notification_type_content_function
 
 FILTER_AUDIT_EXPIRED_USERS_WITH_NO_ROLE = 'filter_audit_expired_users_with_no_role'
 
@@ -458,16 +459,6 @@ class NotificationAppManager:
         return course_notification_preference_config
 
 
-def get_notification_type_content_function(notification_type):
-    """
-        Returns the content function for the given notification if it exists.
-    """
-    try:
-        return globals()[f"get_{notification_type}_notification_content"]
-    except KeyError:
-        return None
-
-
 def get_notification_content(notification_type, context):
     """
     Returns notification content for the given notification type with provided context.
@@ -484,36 +475,6 @@ def get_notification_content(notification_type, context):
         notification_type_content_template = notification_type.get('content_template', None)
         if notification_type_content_template:
             return notification_type_content_template.format(**context, **html_tags_context)
-    return ''
-
-
-def get_new_comment_notification_content(notification_type, context):
-    """
-        Returns notification content for the new_comment notification.
-    """
-    return get_notification_content_with_author_pronoun(notification_type, context)
-
-
-def get_comment_on_followed_post_notification_content(notification_type, context):
-    """
-        Returns notification content for the comment_on_followed_post notification.
-    """
-    return get_notification_content_with_author_pronoun(notification_type, context)
-
-
-def get_notification_content_with_author_pronoun(notification_type, context):
-    """
-        Helper function to get notification content with author's pronoun.
-    """
-    html_tags_context = {
-        'strong': 'strong',
-        'p': 'p',
-    }
-    notification_type_content_template = notification_type.get('content_template', None)
-    if 'author_pronoun' in context:
-        context['author_name'] = context['author_pronoun']
-    if notification_type_content_template:
-        return notification_type_content_template.format(**context, **html_tags_context)
     return ''
